[archivist] Remove unused call to Tree#GetContent
We used to do this so we held a handle to the root VMO of a component's
inspect data so that later we could use it for recording post-mortem
inspect when a component is stopped. However, at the moment we are not
doing using it for anything and it doesn't seem like we will anytime
soon. So for now, let's remove this unnecessary FIDL call that we do
when reading data from most components.

Tested: existing tests

Change-Id: I1aa08f52e8b272ddd2ecb879533e6dd7f5dc64a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569925
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>