[hlcpp] Use correct type shape version in for request allocation
This may have broken out of line objects in compatibility tests.
No test is included in this CL because it is a small fix that
I want to get in quickly, but I am planning to make better
integration testing of the new wire format a focus in the near
term.

Bug: 81995

Change-Id: I8776858b1580cd6f97afa567b958ef9e9f5fc408
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570402
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>