[audio_core] Idle-power on output hot-swap
When an output device is added or removed when streams are actively
playing, IdlePolicy needs to account for any change in routing. Note
that in these cases, the number of active renderer streams does not
change, but an audio output device may no longer be the target of
active streams. This accounting is needed so that IdlePolicy correctly
manages the enabled/disabled states of audio output devices, not just
as AudioRenderer streams come and go, but also as other output devices
come and go.

AudioCore extracts the subset of RouteGraph that is used by
AudioDeviceManager, into the new interface DeviceRouter. IdlePolicy
implements the interface (and is injected into AudioDeviceManager
instead of RouteGraph), using it to take snapshots of the route graph
before and after devices are added/removed. Between taking these
snapshots, it passes the actual AddDeviceToRoutes and
RemoveDeviceFromRoutes calls through to the actual RouteGraph.

Also, debug-logging for the idle-power mechanism is reduced and
cleaned up, now that support for this feature has stabilized across
AudioCore and the audio drivers.

Two top-level build-time flags are added to enable/disable the entire
idle-power feature, so device power consumption can be easily
compared, before and after this feature. These can be removed in a
subsequent CL.

Bug: fxb/75709
Bug: fxb/75714
Bug: fxb/82302
Bug: fxb/82408
Bug: fxb/82423
Bug: fxb/82798

Change-Id: I88ece39daed54e64d5b2ef4ae091bcdb88d44fc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569502
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
(cherry picked from commit c9fe9b15c06c70f6bacfd636ac189cbcd95b80a4)