[fdio] Avoid interpreting uninitialized items as wait results in ppoll
In ppoll, fdio iterates through a set of FDs and computes a wait_many
handle set. This function had a bug where it would not keep track of the
amount of entries in this set correctly and would interpret
uninitialized stack memory as observed kernel signals in the function's
postable. This could result in incorrect state transitions for fdio
objects that update their state based on observed kernel state, such as
stream sockets.

Based on tamird@'s change in
https://fuchsia-review.googlesource.com/c/fuchsia/+/570502/

Bug: b/194705562

Change-Id: I87e0af28663774e21ba0841ecd67786d245b1f3f