[ffx] remove outer mut from generated macro code
The ffx_plugin macro generates a wrapper function which then calls into
the function which is given. That wrapper function does not have the
same mutability requirements as the actual implementation of the plugin.
Therefore you can get into situations where the compiler warns you about
an unused mut in the generated code because you need a mut in your
implementation. This change removes any mut from the command argument to
the wrapper function while preserving it in the implementation.

Change-Id: I4e43814639a87f4a67df7bf121f49f04e3dbfc94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570304
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>