[session] Migrate usages of BindChild.
This CL migrates usages of `fuchsia.sys2.Realm/BindChild`
to `/OpenExposedDir` and `fuchsia.component.Binder`.
The former is used to access an element's exposed
capabilities; the latter is used to start the element.
Note, this places a requirement for all elements to
expose `fuchsia.component.Binder`.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Bug: 81401
Bug: 81400

Change-Id: I75c1045324ee067c26aa3764227897b1d54e7cb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563821
Reviewed-by: Rich Kadel <richkadel@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>