[camera][isp] Add util to sensor mode
Add a command line utility to interact with the camera's
sensor.

The only command available initially is the command to set
the camera sensor's default operating mode in the ISP
driver such that, upon a camera config change, the new mode
will take effect in the sensor.

Bug: b/196969515
Test: Follow src/camera/bin/sensor_cli/README.TXT to test
      both regular and test pattern mode.
Change-Id: I5362a85e2216ebb87f947f980358be91b04a57e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569745
Commit-Queue: Ernest Hua <ernesthua@google.com>
Reviewed-by: John Sasinowski <jsasinowski@google.com>