[hlcpp] Fixes needed to get compatibility test working
These are fixes to two bugs that when fixed result in the FIDL
compatibility tests passing for all compatibility test servers
except for go and dart, which fail some of the tests.

These bugs are oversights that are difficult to cover in a
useful way by non-integration test tests.

Bug: 81995

Change-Id: Iaf980ef9694e049b716297bf4aebcf60fa6bb140
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570861
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>