[fidl][dart] Fixes for dart that make compatibility tests pass
This CL includes fixes for sizes and offsets in dart method decoding
that result in the dart compatibility tests passing.

I am hoping to commit tests that will catch any regressions, but this
will take time and I'm hoping to get the fixes in first because they
need to roll.

Bug: 80880

Change-Id: I5b7a72f7aaf4ddb7573e94a5b478431ec9183637
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570863
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>