[ffx] Implement `ffx component create`
`ffx component create` is the advanced version of `ffx component run`.
Developers specify the precise moniker of the component they wish
to create along with the component URL.

The `create` command will connect to the Realm protocol in the
incoming namespace of the parent and use the protocol to create
a component instance.

Common code for `create` and `run` has been migrated to the common
library.

The laboratory shard's `expose` decl for the Realm protocol has been
changed to a `use` decl. This allows it to also be used with
`ffx component create`.

Test: fx test ffx_component_create
Change-Id: I540a41e88033c96a91c417a122729277d8e2199b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562621
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>