[assembly] Use sdk manifest to find tools
The fvm, minfs, blobfs, and zbi tools are now located using the sdk
manifest. When building in-tree, the manifest points to host tools
available in `out/default/host_x64/<tool>`, but when run from the SDK,
the manifest points to the tools in the SDK.

This commit also refactors how we handle tools in tests. We now have a
helper function which can generate a fake tool at a specified path. The
tools are faked, and the paths to the tools are passed down to their
respective builders for usage.

Bug: 76378, 79072
Change-Id: Ib1163873a614762e2a7f7485f3416a566593958f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570097
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>