[ermine] Enlarged the touch target size of the close button on a button chip
Since the touch target size of the close button on an AppChip was set too tight, it was hard to press it, especially by tapping it with a finger on a touch screen. Now the button has the default touch target size so that it can be easily hit.

Screenshot: https://screenshot.googleplex.com/NsSNWxYmrWJjyh2

Fixed: 82529

Change-Id: I2d2e568d06d3b9419ac8556f9a9ffd85bb98e53c
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/570723
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Charles Whitten <cwhitten@google.com>
Commit-Queue: Yeonhee Lee <yhlee@google.com>