[wlancfg] Runs regulatory manager until failure.
This change allows the regulatory manager to continue
running in the cases where setting the regulatory region
fails and when a country code of an invalid length is
encountered.  The loop will still terminate in the event
that the regulatory manager can no longer receive updates
from the RegulatoryRegionWatcher service.  Such a failure
will be propagated to the wlancfg main.

Bug: 82663
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Change-Id: I1b7e68bda6334cf5b893f8d3c0384e34f8dbce0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568567
Reviewed-by: Naomi Mccracken <nmccracken@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>