[swd][security] Add policy labels import to core.gni
Adds an import to make policy_labels available in core.gni for leaf
product definitions. Does not define a policy_label in core.gni yet.

This is essentially a no-op since in //build/security/policies.gni, the
policy_label arg is an empty dictionary. This is part of step 3 in the
SWD policy migration.

Change-Id: I117d124087476bfc5cae31c4ba0b35849a366284
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/571311
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Daniel Hwang <arkay@google.com>