[ffx] remove unused mut from target flash
It turns out this mut is just really unused and is not a spurious issue
related to the plugin macro. If the implementation was not split out
into a separate function then this would be an issue. With the
implementation being a separate function with its own mut declaration on
the command parameter, you end up not needing the one at the top level
as you are simply moving that value into the called implementation
function.

Change-Id: If6f92a1a0ee7377dce985259cefeca61d9bfeda9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570305
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>