[vboot-fwparam] Add cros-nvtool.
This is a subset of the "crossystem" tool on ChromeOS. The crossystem
tool gathers data from several sources, including nvram, and allows
interaction with them. cros-nvtool is similar, but it only interacts
with nvram.

This is a raw binary that gets installed to bootfs so that we can use it
from zedboot. This CL also adds all the necessary plumbing to give the
various shells (serial and ssh) access to the `fuchsia.vboot.fwparam.FirmwareParam`
protocol.

Bug: 82432
Test: cros_nvtool get && cros_nvtool set dev_boot_usb true; reboot and
      check the settings stuck.
Change-Id: I790e70bccbf0024561d5ac488712601bcc444501
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569203
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>