[ffx] move build_info out of core
The library src/developer/ffx/core/version/libffx_build_version.rlib is
in the critical path of a lot of builds. I am not 100% on why this is
the case, but it seems it is because ffx/core depends on it and ffx/core
is in the critical path.

This makes ffx/core not dependent on this build version library and
instead makes that library stand alone and requires binaries that need
it to depend on it directly. In this case that is just the ffx binary.

We need this information in the daemon so the daemon lib depends on it.
We also use it in a few plugins. To service those plugins I added this
to the Injector so that plugins can request this information be injected
into their implementation. In general plugins that want this info should
be using the build info service. However, we want this information in
the doctor and version plugins both of which we want to use without
access to a device.

I am not sure how much this helps the build critical path, but this is
what it looks like to move the version out of core.

Bug: 82887
Change-Id: I5b07df8162e43adcbe800db2e1f451788c006706
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570858
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Shai Barack <shayba@google.com>