[rust] Fix linking against C++ stdlib in Rust binaries
The 'ldflags' from the existing '_cpp_standard_library' configs don't
get applied to Rust binary builds. Their equivalent in 'rustflags' has
been added to these configs.

The general rustflags in the 'compiler' config also needed to be updated
to add one of the default Clang search paths (which
contains libc++) which was missing. The other search path was incorrect
- it was using the Rust target triple, which is different than Clang's,
and had the 'lib' subdirectory in the wrong place - and was also fixed.

In addition, the default dependency on 'static_cpp_standard_library' in
the rustc_binary template has been removed. Before these fixes, that
default dependency was a no-op, so removing it avoids linking the C++
standard library in all Fuchsia Rust binaries when it isn't actually
required.

Change-Id: I37470b7d13c71c880bc5b6d47fbd4e47dbac619b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569041
Commit-Queue: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>