[rust][ffi] Create a basic example of cxx usage on Fuchsia
This is a modified version of the demo from
https://github.com/dtolnay/cxx, adapted to demonstrate usage of cxx in
the Fuchsia tree and usage of the generated from both C++ and Rust.

This also serves as a basic test of the rust_cxx_ffi_library template
and cxx integration in general.

Change-Id: Id83ff346274735ab6a1899a657d6fd40ea17ab5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562423
Fuchsia-Auto-Submit: Bryan Henry <bryanhenry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>