[tpm] Move command handling to a thread.
This CL moves command handling to a thread in preparation for taking
requests from other applications. Commands are submitted to a queue which
contains the command to send and a callback to call on its completion.
The callback is called with a zx_status_t (in case command transmission
went wrong) and a TpmResponseHeader* (which is the result from the TPM).

The TPM now no longer performs blocking operations during suspend/init
and command responses are passed back to the caller of the command.

Bug: 81433, 82504, 82505
Change-Id: Ia966257fc6b314e6e7ee922c37d5bba58d3acbfc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572722
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>