[utest] Attempt to remove flake in PortTest::CloseQueueRace.
We currently see a hang in the test in CI/CQ roughly once per
month. The race appears to be:

  * Thread A enqueues a message on a port.

  * Thread A decides there are too many messages on the port, and
    attempts to dequeue a message by calling zx_port_wait.

  * Inside the kernel, the handle to the port is converted to
    a reference to the port object, but the queue of messages is not yet
    inspected.

  * Thread B deletes the handle to the port. In the process, all
    messages on the port are deleted.

  * Inside the kernel, thread A determines that no messages are
    on the port queue, so it waits forever for one, causing a hang.

We resolve this by using an infinite-past deadline for the zx_port_wait
operation, and trying to read a message again if a ZX_ERR_TIMED_OUT
error status is returned.

Thanks to maniscalco@ for spotting the cause of the hang.

Tested with 100,000 iterations, 15 of which triggered the
ZX_ERR_TIMED_OUT race.

Bug: 83334
Fixed: 82914
Change-Id: Idb1007d8462316983c5dbb5a30508013086528fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/571623
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>