[kernel][preempt] Clear local preempt pending in int_handler_finish
During the call to int_handler_finish at the end of exception
processing, if preemption is allowed after dropping the preempt
disable count added at the start of the interrupt handler, and local
preemption is enabled, clear the local preempt pending state
immediately, instead of waiting for it to be cleared later on in the
scheduler.

Doing this prevents a window of opportunity in which the system might
perform an extra preempt operation (at best), or suffer from infinite
recursion (at worst), if preemption is disabled and re-enabled again
after the call to int_handler_finish, but before the preemption
operation has been completed.  See fxb/83079 for details.

Fixed: 83079
Tests: verified using repro steps described in the bug.

Change-Id: If008679f5df32d6b0b41b739ed5c621f1e839006
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572604
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>