[identity] Remove redundant Identity API methods.
This is a first in a chain of CLs to adapt the existing identity API
surface to the simple authentication prototype that we're in the process
of building.

Step 1 is to remove the methods that are now out of scope for Fuchsia,
namely deep intregration with cloud based identity services (this can
be added by products if needed) and for a global notion of an
multi-device account (again products could create this if they needed).

Bug: 81605
Change-Id: Ic2f3da6cc5ddb50b59f1a8ef36d62e817d438821
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570825
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>