[cts] Don't verify deps when releasing the SDK.
sdk_atom.gni (the basis for every SDK rule) will now save the filepath
for every file in the active SDK in a metadata field named "cts_artifacts".
We use that metadata field to define the CTS SDK, and infra uploads
the files in that list to CIPD.

When releasing the CTS SDK, we need to make sure the dependency tree
only includes things we want to vend to CIPD. So we have to skip this
verify step in all of the rules.

This is OK because the deps will be validated during regular developer
workflows (when the CTS isn't being released).

Change-Id: I797e97dda39f0659da9b208426b19f9cf31795c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573369
Reviewed-by: John Shamoon <johnshamoon@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>