[scenic] Make the gfx::Scene object trigger on_view_created_
This is a workaround for the root session effectively being capable of
having two views (in the view tree): the scene and the view.
Because the scene object is the only guaranteed singular root of the
scene graph we want to report that as the "view" of the root session.

This CL make it so that the first scene or view object created by a
session is what is marked as the session's view when registering
protocols with CreateSessionT().

In theory this could prevent the root session's non-focus protocols
from working correctly, but all the current ones are useless for the
root session anyway.

Bug: 82643

Change-Id: I58de58b4b8f3b09bfa7400dfc4910024472a3991
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570848
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>