[audio_core] Reduce JamSync frequency
This CL reduces the usage of jam-sync when synchronizing audio clocks.

At the beginning of a mix, we compare the current position to what we
expected it to be (based on saved values from the previous mix). If
the difference is sufficiently large, we unceremoniously "jam" the
source and destination positions to new values, accepting a (usually)
small discontinuity as the price to pay for regaining synch. Four
changes decrease the occurrence and impact of jam-sync operations.

First, we early-exit the clock synchronization process altogether when
no sync is needed (SyncMode::None). This occurs when source and dest
clocks are the same, as well as when they differ but are DeviceFixed
and in the same clock domain. We do still need to set the initial
inter-clock position relationship, but once this is done no jam-sync
will ever be needed (as the rate relationship is constant).

Second, we expand our tolerance to "rollback" dest position, from 29
frames to 480. This too-low const value was treating valid conditions
(whenever MixJob boundaries did not align with effects' processing-
block boundaries) to be ) as dest-position discontinuities, leading to
continual jam-sync on that segment of the audio graph. In the future,
the Mixer will always enforce the invariant that mix time always goes
forward, but in the meantime we will allow backwards movement up to
this newly-increased limit, which not coincidently represents 10 ms at
48kHz (our internal frame rate when processing mixes/effects).

Third, we expand our source position error threshold from 1 to 2.5 ms.
By increasing this value, we try reconciliation approaches like clock
adjustment or micro-SRC for longer, before resorting to discontinuity.
The 2.5 ms value was expressly chosen, as the magnitude of error that
micro-SRC can eliminate in approx 1 second (at its maximum adjustment
factor: 2500 ppm).

Having significantly reduced the incidence of jam-sync, we can now
reenable logging for jam-sync occurrences. This logging is throttled
by stride values (kJamSyncWarningInterval etc.) defined for three
logging levels. Logging is further throttled for clock-adjustments
as well.

Bug: b/196242499

Change-Id: I2ea4508ada3e61ff9f019eaf1fb17f409a39b925
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572624
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>