[kernel][oom_killer]: Fix sort order.
Originally qsort was used which takes a comp function that returns negative zero or positive, depending on the relationship between a and b, on comp(a,b).

That is if (a-b < 0) implies a < b, a- b == 0 implies a == b and a - b > 0 implies a > b.

CPP stable sort takes a comp function that returns true if a < b, to be read a goes before b in the collection.

The implementation was kept but the meaning changed.

qsort comp function was intentionally written like that, which would produce the descending order expected in the loop following the sort.

On the other hand, stable_sort(while keeping the same comp function) produced the ascending order, which breaks the expectations of the loop following the sort.

The test caught a bug, but the way the code was writing relying on an implicit bool to int conversion of a relation operation caused the problem.

Also the problem might have gone unnoticed, since the existing behavior is to reboot on OOM.

The fix is simple, flip the relationship operator.

Bug: 45616 #Done
Test: k ut job
Change-Id: I2a9033a136721584e9488265d7f4be14bcf9b0ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569666
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>