[cts] Run generate.py to produce the finalized CTS SDK
This CL runs generate.py on the CTS SDK, which pulls in the needed
dependencies and adds BUILD.gn files, so we can properly build the tests.

In the future we will want to transform the CTS SDK _after_ we retrieve
it from CIPD, but we can't do that in CQ yet.

Change-Id: I3e427580eb0062adbe0dfc79efea962a2a0367d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573562
Reviewed-by: John Shamoon <johnshamoon@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>