[hlcpp] Cleanup of hlcpp-focused variants of fidl_decode
Currently, it isn't obvious that the *_skip_unknown_handles
variants of fidl_decode are specialized for hlcpp and the
new naming makes this clearer.

In addition, now that it is clearly hlcpp-focused it is
possible to remove unused variants of the functions.

Change-Id: I28478da0640822a60cf47eec1dac8fd40d3505c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572947
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Ian McKellar <ianloic@google.com>