[kernel][job]: Make OOM Killer Test less brittle.
The OOMKiller says it will kill jobs lower in the tree first( that is
those with lowest height), this establishes a dependency on how the
sorting implementation deals with elements at the same height.

Instead of checking the state after each kill request, what we do is
assert on the initial and final state(hierarchy wise) and for each
intemediate step, we just verify that a job at the right height was
killed. This removes the need to deal with any sort of combination of
which job was killed, and just reduce it to there are X jobs at this
height, at each each intermediate step there should be Y jobs at this
height.

Updated Lock acquisition used during test to take LockOrder.

Test: k ut job
Change-Id: Icbaeaa4058364337a20c229d2de737b02b91ae44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570064
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>