[wlan][policy] Remove active scan decider
This change removes the active scan decider concept in favor of
a purely probabilistic active scan decision.

The hidden network probabilities are now updated after passive scans
so that the most up-to-date information can be used to make an active
scan decision. If an active scan occurs, the hidden network
probabilities may be updated with those results too.

This means that, in practice, there is no way for a caller to
deterministically request a passive-only scan while there are saved
networks, because the decision to active scan is based purely on hidden
network probabilities.

This also required some test refactoring and mocking to maniputlate the
hidden network probabilities to ensure scans ARE deterministic for
testing (i.e. setting the probabilities to 0 or 1).

Multiply: wlancfg-tests
Test: fx test wlancfg-tests
Bug: 75668
Change-Id: I797c40c06f8ff110eb3361d4fb5a44f8919c2926
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/557728
Reviewed-by: Hayden Nix <haydennix@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Fuchsia-Auto-Submit: Hayden Nix <haydennix@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>