[fdio] Add inotify tests that run on Linux and Fuchsia
inotify is a Linux interface that Fuchsia supports through fdio. This
adds a number of tests for the fdio implementation, including several
that run on Linux as well as Fuchsia.  This also changes the error for
an empty path to ENOENT to match Linux.

Bug: 60109
Bug: 73151

Change-Id: Ic7eb13dc92a6ba95658f22a338a27e5d5bdf6375
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574048
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>