[kernel][crashlog] Refactor the platform crashlog API
Refactor the platform crashlog API just a bit in preparation for a
change which will allow crashlogs to be rendered directly into
mappable NVRAM regions, eliminating the need for statically (and
sized) rendering buffers.

There are two main changes to take note of here.  First, the largely
C-style API, along with its individual function pointers, have been
replaced with a C++ vtable indirection approach.  The system starts
with a trivial crashlog implementation which does nothing, and this
trivial implementation may be replaced later on once the crashlog
capabilities of the system are understood.

On ARM systems, this is always a mappable RAM based implementation,
while on x64 this might be either a mappable RAM based implementation,
or an EFI capsule based implementation, depending on what is
discovered in the ZBI.

The second change has to do with how the render buffer is accessed.
Previously, the buffer was a statically allocated 4k buffer located in
platform/power.cc, and used in platform_halt.  Now, the buffer is
fetched from the crashlog implementation by platform_halt instead.
Once the crashlog has been rendered into this buffer, platform_halt
"finalizes" the crashlog, allowing the implementation to take the
steps needed to sequester the final version of the crashlog to survive
across a reboot.

Tests: manual testing on NUC and QEMU ARM64.
Change-Id: Ia68d9622d7ca00cb9b47b380e1d191cb7b931831
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568403
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>