[starnix] Fix symlink accounting
Previously, we were resetting the number of symlinks we would follow at
each step of the resolution process. According to LstatELOOPPath, we're
supposed to apply the symlink budget to the entire resolution process.

This separates the notion of the SymlinkMode from the LookupContext. The
SymlinkMode now just records whether we intend to follow the symlink at
the end of the resolution process. The LookupContext records the number
of remaining symlink follows. As path resolution proceeds, the number of
remaining symlink follows in the LookupContext is mutated, which counts
all the symlinks follows from the whole lookup call tree.

Test: StatTest.LstatELOOPPath
Change-Id: I150238c40343ad49b1a16c1b94d15e731830a208
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574056
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>