[kernel][vm] Retain non-terminal accessed bits
This changes accessed harvesting to not remove the non-terminal
accessed bits if we are not planning to free them. The purpose of this
is to avoid destroying information so that a later FreeUnaccessed
action behaves the same, even if there were 1 (or more) Retain
actions in the middle.

Change-Id: I7542aa39f822de216e7b3957249efd1cc5790dc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572109
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>