[carnelian] Support flatland
Add support for the Flatland compositing system.

Added a new view strategy for Flatland, much of which is a copy of the
strategy for scenic. Since the scenic strategy is going away with
scenic, it might not be worth the effort to share the common
implementation. Reviewers are encouraged to suggest otherwise.

This implementation works, although the lack of stability in Flatland
makes it hard to be sure this support is complete. Ten second traces of
rendering do look very solid, though, with no missed frames. Still, it
seems worthwhile to commit this CL as a basis for further efforts on
Flatland support.

Bug: 83510

Testing: manual

Change-Id: Ie72dd28f57f941377f6a6971693ce1be98735493
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569921
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Rob Tsuk <robtsuk@google.com>