[crashlog] Allow 0-copy crashlog stowing.
Update the ram crashlog library in order to give users the ability to
render their crashlog payload directly into their buffer, skipping the
copy of the crashlog payload during a stow operation.  The old mode of
copying a crashlog payload from a side buffer is still supported.

The main purpose of this change it to allow the kernel to directly
render crashlog payloads into ram-mappable crashlogs, making it a bit
easier to use all of the available crashlog buffer without needing to
guess how much side-buffer they will need to have allocated in order
to do so.

Test: updated unit tests.
Change-Id: Ie7e65e6870baa26dcedcc48c24af780084844dab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568404
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>