[crashlog] Render RAM mappable crashlog payloads in-place.
When we are on a system with crashlog NVRAM which is directly mappable
(ARM64 systems for the most part), and a crash happens, render the
crashlog payload directly to the NVRAM region instead of rendering it
to a side buffer and they copying the payload into the crashlog.

While this does save a small amount of memory (we don't need to keep
around a statically allocated side buffer), the primary benefit here
is that we can make use of _all_ of the NVRAM allocated to the
crashlog without needing to guess at how large the statically
allocated side buffer would need to be in order to actually fill the
log.

This used to be a moot point since our crashlogs used to (more or
less) always fit into our 4kB side buffer.  Now that we have the
ability to attempt to render as much of the recent dlog as we can into
the crashlog, however, having the extra space really does help.

Tests: manual testing on NUC and QEMU ARM64
Change-Id: I61e64037ae14bff1bf26679df8cb3947ea9df9d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568405
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>