[ffx][config] Remove RealPaths
For Sdks built from a sdk directory, the prefix was being appended
twice, causing the path to be incorrect. We do not need the Prefix type
for RealPaths, and can replace it with a no-op. And to clean it up even
further we can remove RealPaths altogether and use an optional map.

Bug: 83609
Change-Id: I09bb958edf153a3e11aee2380195db1ad57a0635
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574204
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Christopher Crawford <chriscraws@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>