[wayland] Flatland support.
This implements initial wayland bridge support for Flatland.

A build flag is used to enable Flatland support. This is
required as wayland clients needs to be able to build a
scene prior to a view has been created.

The priority has been to make Flatland code easy to read
and GFX code easy to remove when no longer needed. Some code
duplication was added as a result.

Use statements have been updated to better support this
conditional build.

This also improves the XDG shell implementation by handling
initial commits correctly. This was required as layout
event timing is different enough for Flatland that this
is required.

Bug: 64205
Change-Id: Ie6c0bf729da9de8c6aa5f31f4755482321d8db75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573425
Commit-Queue: David Reveman <reveman@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>