[kernel][hypervisor] Ensure vmclear is run on correct CPU.
AutoPin is a class intended to pin the current thread to a particular
CPU. It is currently only used during Vcpu deletion to ensure that
vmclear is executed on the CPU that last ran the Vcpu.

However, due to a bug, AutoPin doesn't actually migrate the currently
running thread to another CPU. Instead, vmclear is always run on the
current CPU instead of the CPU the VCPU last ran on.

This CL replaces the single use of the AutoPin class with
a mp_sync_exec() call that executes the vmclear on the correct CPU.
Additionally, we add a new test case exercising this path.

Change-Id: Ic953d9845f8705e011b39dd8889e8f50320a369a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573204
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>