[test][cprng] Add more cmdline entropy tests.
This CL adds two more tests that exercise the cmdline entropy behavior.

Basically, one test invokes the emulator twice, using different cmdline
entropy, and runs a binary that prints random numbers. If the cmdline
entropy is being processed correctly, the set of random numbers must be
disjoint.

The other test is similar, except that it uses the same cmdline entropy
on both invocations. The expected result is that the random numbers
overlap at least once. This is used to validate that the first test is
indeed using the cmdline entropy as the only source of entropy.

In order to use only one source of entropy, the CL adds two boot
options: `kernel.cprng-disable.jitterentropy` and
`kernel.cprng-disable.hw_rng`, to disable external sources of entropy.
Note that the global cprng still honors the cprng-seed-required and
cprng-reseed-required flags, meaning that if any of those are set, and
the source of entropy is disabled, the kernel will panic.

These tests assume that nobody else uses the cprng (we only have one
userspace process, and the kernel doesn't use it), and that the uses
that happen before the first userspace process are all the same.

Test: fx test kernel_cprng_tests
Multiply: kernel_cprng_tests : 200
Bug: 82809
Change-Id: I694e74e748821599be95dbac9f3b5c7f77c5c021
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573144
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>