[tools] Move //scripts/codesize to //tools/codesize
This change adds "codesize" to tools, which then makes "codesize"
visible in $root_out_dir/tool_paths.json. This is necessary in order
for the infrastructure to locate the tool.

The tool has been added to //bundles:build_infra so that the
infrastructure builds it by default.

fx codesize workflow is preserved (though it is still broken by recent
changes to blobs.json).

Test: fx set core.arm64 --release --with //tools:tests
      fx build
      fx test //tools/codesize:codesize_tests -vo

Bug: 83452
Change-Id: Ie6774dd48e3e8769ac1ede0bc20e305a3fb06e52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574702
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Anthony Fandrianto <atyfto@google.com>