[wlan][iwlwifi] try to load the next version file if unavailable
This is a follow-up of c1df548d8ce4 where we missed the retry
mechanism.  In this CL, the code will try to load the next version
if the current version file is not available.

Bug: 76121
Tested: On 9260, the max version is 44 but the available firmware
        version is 43.

Change-Id: Ic895707ca02856080a52fb1947ad01aa3b92aec1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562782
Commit-Queue: Louis Yung-chieh Lo <yjlou@google.com>
Reviewed-by: John Sheu <sheu@google.com>