[virtualization][arm64] Load Linux kernels without EFI support.
ARM64-based Linux kernels can be compiled with or without EFI support.
In both cases, they require a header as described in the Linux kernel
docs [1].

When a kernel is compiled with EFI support, it additionally requires
a "MZ" prefix at the beginning of the header, and some additional
reserved fields filled in, as described in [2].

The VMM doesn't require EFI support to be compiled into the kernel, and
can happily load kernels without it. Remove the MZ header checks.

[1]: https://www.kernel.org/doc/Documentation/arm64/booting.txt
[2]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/Documentation/arm64/booting.txt?id=4370eec05a887b0cd4392cd5dc5b2713174745c0

Bug: 82425
Change-Id: Ia1e687317e5de0072a45327041d04032ffa5bf10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574812
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>