[fidl][llcpp] Move generated bindings
This moves the generated bindings from:
  #include <library/name/llcpp/fidl.h>
  #include <library/name/llcpp/fidl_test_base.h>
to:
  #include <fidl/library.name/cpp/wire.h>
  #include <fidl/library.name/cpp/wire_test_base.h>
and the fidl.cc to wire.cc too.

The plan is for the unified C++ bindings to all reside in this same
location.

This CL includes a forwarding header for the old llcpp/fidl.h location
but and the llcpp/fidl_test_base.h. The use sites will be updated and
the forwarding headers removed in future CLs.

Bug: 74067
Change-Id: I3d1ccd80023a7521aa3507f939c0948d6a8fd0f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573485
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>