[fidl] Add a comment to clarify light sensor data lux value to be 'clear'.
And delete old comment since it's wrong in current Nest light sensor chips.
The lux field from the light sensor is actually the clear value, it is not lux. Any use of the lux value has to be used as the clear value. It is incorrectly named.
Possibly deprecate and make new field with correct name in the future.

Change-Id: I6baa9361110cd3a666978da81fd60e2b7fe0af51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/572525
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Ian Gonzalez <iangonzalez@google.com>
Commit-Queue: Fengge Liu <fenggeliu@google.com>
API-Review: Kevin Lindkvist <lindkvist@google.com>